kvutils refactoring
Move transaction submission processing to its own class and refactor to
stop early on errors.

Throw a known error rather than use sys.error.
